Support for two finger interactions
