+++ packages/viewer/src/interactions/pointerInteractionHandler.ts
@@ -16,15 +16,4 @@ export class PointerInteractionHandler extends BaseInteractionHandler {
       new PanInteraction()
     );
   }
-
-  protected registerEvents(element: HTMLElement): void {
-    element.addEventListener('pointerdown', this.handleDownEvent);
-    element.addEventListener('wheel', this.handleMouseWheel);
-  }
-
-  protected deregisterEvents(): void {
-    this.element?.removeEventListener('pointerdown', this.handleDownEvent);
-    this.element?.removeEventListener('wheel', this.handleMouseWheel);
-    this.element = undefined;
-  }
 }
